Group 14 All issues solved

+++ src/argouml-app/src/org/argouml/cognitive/ToDoList.java
@@ -195,11 +195,11 @@ public class ToDoList extends Observable implements Runnable {
 
             forceValidityCheck(removes);
             removes.clear();
-            try {
-                Thread.sleep(SLEEP_SECONDS * 1000);
-            } catch (InterruptedException ignore) {
-                LOG.log(Level.SEVERE, "InterruptedException!!!", ignore);
-            }
+           try {
+    Thread.sleep((long) SLEEP_SECONDS * 1000);
+} catch (InterruptedException ignore) {
+    LOG.log(Level.SEVERE, "InterruptedException!!!", ignore);
+}
         }
     }
